testutil: support go 1.13 change to testing/flag package usage
