+++ api/api_test.go
@@ -21,7 +21,6 @@ import (
 	"context"
 	crand "crypto/rand"
 	"errors"
-	"flag"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -40,9 +39,7 @@ import (
 )
 
 func init() {
-	loglevel := flag.Int("loglevel", 2, "loglevel")
-	flag.Parse()
-	log.Root().SetHandler(log.CallerFileHandler(log.LvlFilterHandler(log.Lvl(*loglevel), log.StreamHandler(os.Stderr, log.TerminalFormat(true)))))
+	testutil.Init()
 }
 
 func testAPI(t *testing.T, f func(*API, *chunk.Tags, bool)) {
@@ -143,7 +140,7 @@ func TestApiPut(t *testing.T) {
 		resp := testGet(t, api, addr.Hex(), "")
 		checkResponse(t, resp, exp)
 		tag := tags.All()[0]
-		testutil.CheckTag(t, tag, 2, 2, 0, 2) //1 chunk data, 1 chunk manifest
+		chunk.CheckTag(t, tag, 2, 2, 0, 2) //1 chunk data, 1 chunk manifest
 	})
 }
 
@@ -170,11 +167,11 @@ func TestApiTagLarge(t *testing.T) {
 		if toEncrypt {
 			tag := tags.All()[0]
 			expect := int64(4095 + 64 + 1)
-			testutil.CheckTag(t, tag, expect, expect, 0, expect)
+			chunk.CheckTag(t, tag, expect, expect, 0, expect)
 		} else {
 			tag := tags.All()[0]
 			expect := int64(4095 + 32 + 1)
-			testutil.CheckTag(t, tag, expect, expect, 0, expect)
+			chunk.CheckTag(t, tag, expect, expect, 0, expect)
 		}
 	})
 }
