+++ cmd/swarm/upload_test.go
@@ -33,12 +33,10 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	swarmapi "github.com/ethersphere/swarm/api/client"
 	"github.com/ethersphere/swarm/testutil"
-	"github.com/mattn/go-colorable"
 )
 
 func init() {
-	log.PrintOrigins(true)
-	log.Root().SetHandler(log.LvlFilterHandler(log.Lvl(*loglevel), log.StreamHandler(colorable.NewColorableStderr(), log.TerminalFormat(true))))
+	testutil.Init()
 }
 
 func TestSwarmUp(t *testing.T) {
