+++ pss/pss_test.go
@@ -23,7 +23,6 @@ import (
 	"encoding/binary"
 	"encoding/hex"
 	"encoding/json"
-	"flag"
 	"fmt"
 	"io/ioutil"
 	"math/rand"
@@ -49,12 +48,11 @@ import (
 	"github.com/ethersphere/swarm/p2p/protocols"
 	"github.com/ethersphere/swarm/pot"
 	"github.com/ethersphere/swarm/state"
+	"github.com/ethersphere/swarm/testutil"
 )
 
 var (
 	initOnce        = sync.Once{}
-	loglevel        = flag.Int("loglevel", 2, "logging verbosity")
-	longrunning     = flag.Bool("longrunning", false, "do run long-running tests")
 	w               *whisper.Whisper
 	wapi            *whisper.PublicWhisperAPI
 	psslogmain      log.Logger
@@ -66,7 +64,7 @@ var (
 )
 
 func init() {
-	flag.Parse()
+	testutil.Init()
 	rand.Seed(time.Now().Unix())
 
 	adapters.RegisterServices(newServices(false))
@@ -77,10 +75,6 @@ func initTest() {
 	initOnce.Do(
 		func() {
 			psslogmain = log.New("psslog", "*")
-			hs := log.StreamHandler(os.Stderr, log.TerminalFormat(true))
-			hf := log.LvlFilterHandler(log.Lvl(*loglevel), hs)
-			h := log.CallerFileHandler(hf)
-			log.Root().SetHandler(h)
 
 			w = whisper.New(&whisper.DefaultConfig)
 			wapi = whisper.NewPublicWhisperAPI(w)
@@ -1230,7 +1224,7 @@ func TestNetwork(t *testing.T) {
 // nodes/recipientAddresses/addrbytes/adaptertype
 // if adaptertype is exec uses execadapter, simadapter otherwise
 func TestNetwork2000(t *testing.T) {
-	if !*longrunning {
+	if !*testutil.Longrunning {
 		t.Skip("run with --longrunning flag to run extensive network tests")
 	}
 	t.Run("3/2000/4/sim", testNetwork)
@@ -1240,7 +1234,7 @@ func TestNetwork2000(t *testing.T) {
 }
 
 func TestNetwork5000(t *testing.T) {
-	if !*longrunning {
+	if !*testutil.Longrunning {
 		t.Skip("run with --longrunning flag to run extensive network tests")
 	}
 	t.Run("3/5000/4/sim", testNetwork)
@@ -1250,7 +1244,7 @@ func TestNetwork5000(t *testing.T) {
 }
 
 func TestNetwork10000(t *testing.T) {
-	if !*longrunning {
+	if !*testutil.Longrunning {
 		t.Skip("run with --longrunning flag to run extensive network tests")
 	}
 	t.Run("3/10000/4/sim", testNetwork)
